src: resolve compiler warnings
On i386, gcc warns:

common/ECConfigImpl.cpp: In member function "void ECConfigImpl::InsertOrReplace(settingmap_t*, const settingkey_t&, const char*, bool)":
common/ECConfigImpl.cpp:250:53: warning: format "%lu" expects argument of type "long unsigned int", but argument 4 has type "size_t {aka unsigned int}" [-Wformat=]
   len = snprintf(data, 1024, "%lu", GetSize(szValue));
common/StatsClient.cpp: In member function "void StatsClient::submit(const string&, time_t, int64_t)":
common/StatsClient.cpp:170:82: warning: format "%ld" expects argument of type "long int", but argument 6 has type "int64_t {aka long long int}" [-Wformat=]
  int len = snprintf(msg, sizeof msg, "ADD int %s %ld %ld", key.c_str(), ts, value);
provider/libserver/ECCacheManager.cpp: In member function "ECRESULT ECCacheManager::GetObjects(const std::__cxx11::list<
sObjectTableKey>&, std::map<sObjectTableKey, ECsObjects>&)":
provider/libserver/ECCacheManager.cpp:38:77: warning: format "%lu" expects argument of type "long unsigned int", but argument 3 has type "std::__cxx11::list<sObjectTableKey>::size_type {aka unsigned int}" [-Wformat=]
  ec_log(EC_LOGLEVEL_DEBUG | EC_LOGLEVEL_CACHE, "cache: " _msg, ##__VA_ARGS__)
provider/libserver/ECCacheManager.cpp:497:3: note: in expansion of macro "LOG_CACHE_DEBUG"
   LOG_CACHE_DEBUG("Get object ids total ids %lu from disk %lu", lstObjects.size(), setUncached.size());
provider/libserver/ECCacheManager.cpp:841:2: note: in expansion of macro "LOG_USERCACHE_DEBUG"
  LOG_USERCACHE_DEBUG("Get User Objects. requested objects %lu", lstExternObjIds.size() );
provider/libserver/ECCacheManager.cpp:907:3: note: in expansion of macro "LOG_USERCACHE_DEBUG"
   LOG_USERCACHE_DEBUG("Get User Objects done. Returned objects %lu", lpmapLocalObjIds->size());
ECtools/archiver/ArchiveStateUpdater.cpp: In member function "HRESULT ArchiveStateUpdater::RemoveImplicit(const entryid_t&, const tstring&, const abentryid_t&, const ObjectEntryList&)":
ECtools/archiver/ArchiveStateUpdater.cpp:289:115: warning: format "%lu" expects argument of type "long unsigned int", but argument 4 has type "std::__cxx11::list<SObjectEntry>::size_type {aka unsigned int}" [-Wformat=]
  m_lpLogger->Log(EC_LOGLEVEL_DEBUG, "Processing %lu archives for implicitly attached archives", lstArchives.size());
ECtools/archiver/ArchiveStateUpdater.cpp: In member function "HRESULT ArchiveStateUpdater::AddServerBased(const tstring&, const abentryid_t&, const std::__cxx11::list<std::__cxx11::basic_string<wchar_t> >&, unsigned int)":
ECtools/archiver/ArchiveStateUpdater.cpp:477:79: warning: format "%lu" expects argument of type "long unsigned int", but argument 4 has type "std::__cxx11::list<std::__cxx11::basic_string<wchar_t> >::size_type {aka unsigned int}" [-Wformat=]
  m_lpLogger->Log(EC_LOGLEVEL_DEBUG, "Attaching %lu servers", lstServers.size());
ECtools/archiver/ArchiveControlImpl.cpp:304:122: warning: format "%lu" expects argument of type "long unsigned int", but argument 4 has type "std::__cxx11::list<std::__cxx11::basic_string<wchar_t> >::size_type {aka unsigned int}" [-Wformat=]
  m_lpLogger->Log(EC_LOGLEVEL_INFO, "Processing " SIZE_T_PRINTF "%s users.", lstUsers.size(), (bLocalOnly ? " local" : ""));
ECtools/archiver/ArchiveControlImpl.cpp: In member function "HRESULT ArchiveControlImpl::PurgeArchiveFolder(MsgStorePtr&, const entryid_t&, LPSRestriction)":
ECtools/archiver/ArchiveControlImpl.cpp:927:97: warning: format "%lu" expects argument of type "long unsigned int", but argument 4 has type "std::__cxx11::list<entryid_t>::size_type {aka unsigned int}" [-Wformat=]
  m_lpLogger->Log(EC_LOGLEVEL_INFO, "Purging %lu messaged from archive folder", lstEntries.size());
provider/client/ECPropertyEntry.cpp:644:55: warning: left operand of comma operator has no effect [-Wunused-value]
    assert(("We should never have PT_STRING8 storage", 0));
provider/client/ECPropertyEntry.cpp:680:58: warning: left operand of comma operator has no effect [-Wunused-value]
    assert(("We should never have PT_MV_STRING8 storage", 0));
provider/client/ECPropertyEntry.cpp: In member function "HRESULT ECProperty::CopyTo(LPSPropValue, void*, ULONG)":
provider/client/ECPropertyEntry.cpp:773:54: warning: left operand of comma operator has no effect [-Wunused-value]
   assert(("We should never have PT_STRING8 storage", 0));
provider/client/ECPropertyEntry.cpp:928:57: warning: left operand of comma operator has no effect [-Wunused-value]
   assert(("We should never have PT_MV_STRING8 storage", 0));
ECtools/mapitime.cpp: In function "int mpt_option_parse(int, char**)":
ECtools/mapitime.cpp:235:10: warning: ISO C++ forbids converting a string constant to "char*" [-Wpedantic]
   user = "foo";
gateway/IMAP.cpp: In member function "HRESULT IMAP::HrCmdAuthenticate(const string&, std::__cxx11::string, const string&)":
gateway/IMAP.cpp:843:122: warning: format "%lu" expects argument of type "long unsigned int", but argument 4 has type "std::vector<std::__cxx11::basic_string<char> >::size_type {aka unsigned int}" [-Wformat=]
   lpLogger->Log(EC_LOGLEVEL_INFO, "Invalid authentication data received, expected 3 items, have %lu items.", vAuth.size());
caldav/CalDavProto.cpp: In member function "HRESULT CalDAV::HrHandleFreebusy(ICalToMapi*)":
./common/include/kopano/ECLogger.h:333:66: warning: format "%lu" expects argument of type "long unsigned int", but argument 3 has type "std::__cxx11::list<std::__cxx11::basic_string<char> >::size_type {aka unsigned int}" [-Wformat=]
 #define ec_log_err(...)     ec_log(EC_LOGLEVEL_ERROR, __VA_ARGS__)
caldav/CalDavProto.cpp:1857:3: note: in expansion of macro "ec_log_err"
   ec_log_err("Unable to get freebusy information for %lu users: 0x%08X", lstUsers->size(), hr);
spooler/Spooler.cpp: In function "void process_signal(int)":
spooler/Spooler.cpp:910:80: warning: format "%lu" expects argument of type "long unsigned int", but argument 4 has type "std::map<int, _SendData>::size_type {aka unsigned int}" [-Wformat=]
   g_lpLogger->Log(EC_LOGLEVEL_DEBUG, "Running threads: %lu", mapSendData.size());
spooler/Spooler.cpp:912:81: warning: format "%lu" expects argument of type "long unsigned int", but argument 4 has type "std::map<int, int>::size_type {aka unsigned int}" [-Wformat=]
   g_lpLogger->Log(EC_LOGLEVEL_DEBUG, "Finished threads: %lu", mapFinished.size());

+++ common/StatsClient.cpp
@@ -167,7 +167,9 @@ void StatsClient::submit(const std::string & key, const time_t ts, const int64_t
 		return;
 
 	char msg[4096];
-	int len = snprintf(msg, sizeof msg, "ADD int %s %ld %ld", key.c_str(), ts, value); 
+	int len = snprintf(msg, sizeof msg, "ADD int %s %ld %zd",
+	          key.c_str(), static_cast<long>(ts),
+	          static_cast<size_t>(value));
 
 	// in theory snprintf can return -1
 	if (len > 0) {
